+++ pg_upsert/__main__.py
@@ -1,0 +1,6 @@
+# my_library/__main__.py
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
