+++ pg_upsert/pg_upsert.py
@@ -2,549 +2,23 @@
 
 from __future__ import annotations
 
-import argparse
-import getpass
 import logging
-import re
 import sys
-import tkinter as tk
-import tkinter.font as tkfont
-import tkinter.ttk as ttk
 from datetime import datetime
-from pathlib import Path
-from urllib.parse import urlparse, urlunparse
 
 import psycopg2
-from psycopg2.extras import DictCursor
 from psycopg2.sql import SQL, Composable, Identifier, Literal
 from tabulate import tabulate
 
 from .__version__ import __description__, __version__
+from .database import PostgresDB
+from .ui import (
+    CompareUI,
+    TableUI,
+)
+from .utils import ellapsed_time
 
 logger = logging.getLogger(__name__)
-
-
-class PostgresDB:
-    """Base database object for connecting and executing SQL queries on a PostgreSQL database.
-
-    Args:
-        conn (psycopg2.extensions.connection, optional): An existing connection object to a PostgreSQL database.
-        uri (str, optional): A connection URI for a PostgreSQL database.
-        encoding (str, optional): The encoding to use for the database connection.
-        **kwargs: Additional keyword arguments passed to `psycopg2.connect()`.
-
-    Returns:
-        PostgresDB: A new PostgresDB object for connecting to a PostgreSQL database and executing queries.
-
-    Raises:
-        AttributeError: If neither a connection URI nor an existing connection object is provided.
-        psycopg2.Error: If an error occurs while connecting to the database or executing a query.
-    """
-
-    def __init__(
-        self,
-        uri: None | str = None,
-        conn: None | psycopg2.extensions.connection = None,
-        encoding: str = "utf-8",
-        **kwargs,
-    ):
-        if conn is None and uri is None:
-            raise AttributeError(
-                "Either a connection URI or an existing connection object must be provided.",
-            )
-        if conn and uri:
-            logger.warning(
-                "Connection URI ignored as an existing connection object is provided.",
-            )
-        # If a URI is supplied, check for a password and prompt if necessary
-        if uri:
-            uri = self._prompt_for_password(uri)
-        self.conn = conn or psycopg2.connect(uri, **kwargs)
-        self.encoding = encoding
-        self.in_transaction = False
-        if not self._is_valid_connection():
-            raise psycopg2.Error(f"Error connecting to {self.conn.dsn}")
-
-    def __repr__(self) -> str:
-        """Return a string representation of the object."""
-        params = self.conn.get_dsn_parameters() if self.conn else "No connection"
-        return f"{self.__class__.__name__}({params})"
-
-    def __del__(self):
-        """Ensure the database connection is closed when the object is deleted, if open."""
-        if hasattr(self, "conn") and self.conn and not self.conn.closed:
-            self.close()
-
-    def _prompt_for_password(self, uri: str) -> str:
-        """Prompt the user for a password."""
-        try:
-            parsed_uri = urlparse(uri)
-            if parsed_uri.password:
-                return uri
-            prompt = f"The script {Path(__file__).name} wants the password for PostgresDB(uri={uri}): "
-            return urlunparse(
-                parsed_uri._replace(
-                    netloc=f"{parsed_uri.username}:{getpass.getpass(prompt)}@{parsed_uri.hostname}:{parsed_uri.port}",
-                ),
-            )
-        except (KeyboardInterrupt, EOFError) as err:
-            raise err
-
-    def _is_valid_connection(self) -> bool:
-        """Check if the database connection is valid."""
-        try:
-            with self.conn.cursor():
-                self.conn.set_client_encoding(self.encoding)
-            return True
-        except psycopg2.Error:
-            return False
-
-    def open_db(self) -> None:
-        """Ensure the database connection is open."""
-        if not self.conn or self.conn.closed:
-            logger.debug("Opening database connection.")
-            self.conn = psycopg2.connect(self.conn.dsn)
-            self.conn.set_client_encoding(self.encoding)
-            self.conn.set_session(autocommit=False)
-        elif self.conn.closed:
-            logger.warning("Connection is closed; attempting to reopen.")
-            self.conn = psycopg2.connect(self.conn.dsn)
-            self.conn.set_client_encoding(self.encoding)
-
-    def cursor(self):
-        """Return a cursor for executing database queries."""
-        self.open_db()
-        return self.conn.cursor(cursor_factory=DictCursor)
-
-    def close(self) -> None:
-        """Close the database connection if open."""
-        if self.conn and not self.conn.closed:
-            self.rollback()
-            self.conn.close()
-
-    def commit(self) -> None:
-        """Commit the current transaction."""
-        if self.conn and self.in_transaction:
-            self.conn.commit()
-            self.in_transaction = False
-
-    def rollback(self) -> None:
-        """Rollback the current transaction."""
-        if self.conn and self.in_transaction:
-            self.conn.rollback()
-            self.in_transaction = False
-
-    def execute(
-        self: PostgresDB,
-        sql: str | Composable,
-        params=None,
-    ) -> psycopg2.extensions.cursor:
-        """A shortcut to self.cursor().execute() that handles encoding.
-
-        Handles insert, updates, deletes
-
-        Args:
-            sql (str | psycopg2.sql.Composable): The SQL query to execute. Accepts a `str` or `Composable` object.
-            params (tuple, optional): A tuple of parameters to pass to the query.
-                Note that a `Composable` object should not have parameters passed separately. Default is None.
-        Returns:
-            psycopg2.extensions.cursor: A cursor object for the executed query.
-        """
-        self.in_transaction = True
-        try:
-            curs = self.cursor()
-            if isinstance(sql, Composable):
-                logger.debug(f"\n{sql.as_string(curs)}")
-                curs.execute(sql)
-            else:
-                if params is None:
-                    logger.debug(f"\n{sql}")
-                    curs.execute(sql.encode(self.encoding))
-                else:
-                    logger.debug(f"\nSQL:\n{sql}\nParameters:\n{params}")
-                    curs.execute(sql.encode(self.encoding), params)
-        except Exception:
-            self.rollback()
-            raise
-        return curs
-
-    def rowdict(self: PostgresDB, sql: str | Composable, params=None) -> tuple:
-        """Convert a cursor object to an iterable that yields dictionaries of row data.
-
-        yields dictionaries of row data with the following structure:
-            0) dict_row (iterator) - an iterator that yields dictionaries of row data
-            1) headers (list) - a list of column names
-            2) rowcount (int) - the number of rows returned by the query
-        """
-        curs = self.execute(sql, params)
-        if not curs.description:
-            # No data returned
-            return (iter([]), [], 0)
-        headers = [d[0] for d in curs.description]
-
-        def dict_row():
-            """Convert a data row to a dictionary."""
-            row = curs.fetchone()
-            if row:
-                if self.encoding:
-                    r = [(c.decode(self.encoding, "backslashreplace") if isinstance(c, bytes) else c) for c in row]
-                else:
-                    r = row
-                return dict(zip(headers, r, strict=True))
-            return None
-
-        return (iter(dict_row, None), headers, curs.rowcount)
-
-
-class CompareUI:
-    def __init__(
-        self: CompareUI,
-        title: str,
-        message: str,
-        button_list: list[tuple[str, int, str]],
-        headers1: list[str],
-        rows1: list | tuple,
-        headers2: list[str],
-        rows2: list | tuple,
-        keylist: list[str],
-        selected_button=0,
-        sidebyside=False,
-    ) -> None:
-        """button_list: list of 3-tuples where the first item is the button label,
-        the second item is the button's value, and the third (optional) value is the key
-        to bind to the button.
-        Key identifiers must be in the form taken by the Tk bind() function,
-        e.g., "<Return>" and "<Escape>" for those keys, respectively.
-        keylist: list of column names that make up a common key for both tables.
-        selected_button: integer identifying which button should get the focus (0-based)
-        """
-        self.headers1 = headers1
-        self.rows1 = rows1
-        self.headers2 = headers2
-        self.rows2 = rows2
-        self.keylist = keylist
-        self.return_value = None
-        self.button_value = None
-        self.win = tk.Toplevel()
-        self.win.title(title)
-
-        def hl_unmatched(*args):
-            """Highlight all rows in both tables that are not matched in the
-            other table.
-
-            Create a list of lists of key values for table1.
-            """
-            keyvals1 = []
-            tblitems = self.tbl1.get_children()
-            for row_item in tblitems:
-                rowdict = dict(
-                    zip(self.headers1, self.tbl1.item(row_item)["values"], strict=True),
-                )
-                keyvals1.append([rowdict[k] for k in self.keylist])
-            # Create a list of lists of key values for table2.
-            keyvals2 = []
-            tblitems = self.tbl2.get_children()
-            for row_item in tblitems:
-                rowdict = dict(
-                    zip(self.headers2, self.tbl2.item(row_item)["values"], strict=True),
-                )
-                keyvals2.append([rowdict[k] for k in self.keylist])
-            # Create a list of only unique key values in common.
-            keyvals = []
-            for vals in keyvals1:
-                if vals in keyvals2 and vals not in keyvals:
-                    keyvals.append(vals)
-            # Highlight rows in table 1
-            tblitems = self.tbl1.get_children()
-            for row_item in tblitems:
-                self.tbl1.selection_remove(row_item)
-                rowdict = dict(
-                    zip(self.headers1, self.tbl1.item(row_item)["values"], strict=True),
-                )
-                rowkeys = [rowdict[k] for k in self.keylist]
-                if rowkeys not in keyvals:
-                    self.tbl1.selection_add(row_item)
-            # Highlight rows in table 2
-            tblitems = self.tbl2.get_children()
-            for row_item in tblitems:
-                self.tbl2.selection_remove(row_item)
-                rowdict = dict(
-                    zip(self.headers2, self.tbl2.item(row_item)["values"], strict=True),
-                )
-                rowkeys = [rowdict[k] for k in self.keylist]
-                if rowkeys not in keyvals:
-                    self.tbl2.selection_add(row_item)
-
-        # The checkbox variable controlling highlighting.
-        self.hl_both_var = tk.IntVar()
-        self.hl_both_var.set(0)
-        controlframe = ttk.Frame(master=self.win, padding="3 3 3 3")
-        unmatch_btn = ttk.Button(
-            controlframe,
-            text="Show mismatches",
-            command=hl_unmatched,
-        )
-        unmatch_btn.grid(column=0, row=1, sticky=tk.W)
-        self.msg_label = None
-        # The ttk.Treeview widget that displays table 1.
-        self.tbl1 = None
-        # The ttk.Treeview widget that displays table 2.
-        self.tbl2 = None
-        # A list of ttk.Button objects
-        self.buttons = []
-        self.focus_button = selected_button
-        self.button_clicked_value = None
-
-        def wrap_msg(event: tk.Event):
-            """Wrap the message text to fit the window width."""
-            self.msg_label.configure(wraplength=event.width - 5)
-
-        # Message frame and control.
-        msgframe = ttk.Frame(master=self.win, padding="3 3 3 3")
-        self.msg_label = ttk.Label(msgframe, text=message)
-        self.msg_label.bind("<Configure>", wrap_msg)
-        self.msg_label.grid(column=0, row=0, sticky=tk.EW)
-        # Bottom button frame
-        btnframe = ttk.Frame(master=self.win, padding="3 3 3 3")
-
-        def find_match(
-            from_table: ttk.Treeview,
-            from_headers: list | tuple,
-            to_table: ttk.Treeview,
-            to_headers: list | tuple,
-        ):
-            """Highlight all rows in to_table that matches the selected row in
-            from_table.  Tables are TreeView objects.
-            """
-            sel_item = from_table.focus()
-            if sel_item is not None and sel_item != "":
-                sel_dict = dict(
-                    zip(from_headers, from_table.item(sel_item)["values"], strict=True),
-                )
-                key_dict = {k: sel_dict[k] for k in self.keylist}
-                # Find the matching data (if any) in to_table.
-                to_items = to_table.get_children()
-                to_item = ""
-                found = False
-                for to_item in to_items:
-                    to_table.selection_remove(to_item)
-                    to_dict = dict(
-                        zip(to_headers, to_table.item(to_item)["values"], strict=True),
-                    )
-                    if all(to_dict[k] == key_dict[k] for k in key_dict):
-                        if not found:
-                            to_table.see(to_item)
-                            found = True
-                        to_table.selection_add(to_item)
-
-        def match2to1(event: tk.Event):
-            """Find the matching row in table 1 for the selected row in table 2."""
-            find_match(self.tbl1, self.headers1, self.tbl2, self.headers2)
-            if self.hl_both_var.get() == 1:
-                find_match(self.tbl1, self.headers1, self.tbl1, self.headers1)
-
-        def match1to2(event: tk.Event):
-            """Find the matching row in table 2 for the selected row in table 1."""
-            find_match(self.tbl2, self.headers2, self.tbl1, self.headers1)
-            if self.hl_both_var.get() == 1:
-                find_match(self.tbl2, self.headers2, self.tbl2, self.headers2)
-
-        # Create data tables.
-        self.tablemaster = ttk.Frame(master=self.win)
-        self.tableframe1, self.tbl1 = treeview_table(
-            self.tablemaster,
-            self.rows1,
-            self.headers1,
-            "browse",
-        )
-        self.tableframe2, self.tbl2 = treeview_table(
-            self.tablemaster,
-            self.rows2,
-            self.headers2,
-            "browse",
-        )
-        self.tbl1.bind("<ButtonRelease-1>", match2to1)
-        self.tbl2.bind("<ButtonRelease-1>", match1to2)
-        # Put the frames and other widgets in place.
-        msgframe.grid(column=0, row=0, sticky=tk.EW)
-        controlframe.grid(column=0, row=1, sticky=tk.EW)
-        self.tablemaster.grid(column=0, row=2, sticky=tk.NSEW)
-        if sidebyside:
-            self.tableframe1.grid(column=0, row=0, sticky=tk.NSEW)
-            self.tableframe2.grid(column=1, row=0, sticky=tk.NSEW)
-            self.tablemaster.rowconfigure(0, weight=1)
-            self.tablemaster.columnconfigure(0, weight=1)
-            self.tablemaster.columnconfigure(1, weight=1)
-        else:
-            self.tableframe1.grid(column=0, row=0, sticky=tk.NSEW)
-            self.tableframe2.grid(column=0, row=1, sticky=tk.NSEW)
-            self.tablemaster.columnconfigure(0, weight=1)
-            self.tablemaster.rowconfigure(0, weight=1)
-            self.tablemaster.rowconfigure(1, weight=1)
-        # Create buttons.
-        btnframe.grid(column=0, row=3, sticky=tk.E)
-        for colno, btn_spec in enumerate(button_list):
-            btn_action = ClickSet(self, btn_spec[1]).click
-            btn = ttk.Button(btnframe, text=btn_spec[0], command=btn_action)
-            if btn_spec[2] is not None:
-                self.win.bind(btn_spec[2], btn_action)
-            self.buttons.append(btn)
-            btn.grid(column=colno, row=0, sticky=tk.E, padx=3)
-        # Allow resizing.
-        self.win.columnconfigure(0, weight=1)
-        self.win.rowconfigure(1, weight=0)
-        self.win.rowconfigure(2, weight=2)
-        msgframe.columnconfigure(0, weight=1)
-        btnframe.columnconfigure(0, weight=1)
-        # Other key bindings
-        self.win.protocol("WM_DELETE_WINDOW", self.cancel)
-        # Position window.
-        self.win.update_idletasks()
-        m = re.match(r"(\d+)x(\d+)\+(-?\d+)\+(-?\d+)", self.win.geometry())
-        if m is not None:
-            wwd = int(m.group(1))
-            wht = int(m.group(2))
-            swd = self.win.winfo_screenwidth()
-            sht = self.win.winfo_screenheight()
-            xpos = (swd / 2) - (wwd / 2)
-            ypos = (sht / 2) - (wht / 2)
-            self.win.geometry("%dx%d+%d+%d" % (wwd, wht, xpos, ypos))
-        # Limit resizing
-        self.win.minsize(width=300, height=0)
-
-    def cancel(self: CompareUI):
-        """Cancel the dialog."""
-        self.dialog_canceled = True
-        self.win.destroy()
-
-    def activate(self: CompareUI):
-        """Activate the dialog."""
-        # Window control
-        self.win.grab_set()
-        self.win._root().withdraw()
-        self.win.focus_force()
-        if self.focus_button:
-            self.buttons[self.focus_button].focus()
-        self.win.wait_window(self.win)
-        self.win.update_idletasks()
-        # Explicitly delete the Tkinter variable to suppress Tkinter error message.
-        self.hl_both_var = None
-        rv = self.return_value
-        return (self.button_clicked_value, rv)
-
-
-class TableUI:
-    """A class for displaying a single table in a Tkinter window."""
-
-    def __init__(
-        self: TableUI,
-        title: str,
-        message: str,
-        button_list: list[tuple[str, int, str]],
-        headers: list[str],
-        rows: list | tuple,
-        selected_button=0,
-    ) -> None:
-        self.headers = headers
-        self.rows = rows
-        self.return_value = None
-        self.button_value = None
-        self.win = tk.Toplevel()
-        self.win.title(title)
-        self.msg_label = None
-        self.tbl = None  # The ttk.Treeview widget that displays table 1.
-        self.buttons = []  # A list of ttk.Button objects
-        self.focus_button = selected_button
-        self.button_clicked_value = None
-
-        def wrap_msg(event: tk.Event):
-            """Wrap the message text to fit the window width."""
-            self.msg_label.configure(wraplength=event.width - 5)
-
-        # Message frame and control.
-        msgframe = ttk.Frame(master=self.win, padding="3 3 3 3")
-        self.msg_label = ttk.Label(msgframe, text=message)
-        self.msg_label.bind("<Configure>", wrap_msg)
-        self.msg_label.grid(column=0, row=0, sticky=tk.EW)
-        # Bottom button frame
-        btnframe = ttk.Frame(master=self.win, padding="3 3 3 3")
-        # Create data tables.
-        self.tablemaster = ttk.Frame(master=self.win)
-        self.tableframe, self.tbl1 = treeview_table(
-            self.tablemaster,
-            self.rows,
-            self.headers,
-            "browse",
-        )
-        # Put the frames and other widgets in place.
-        msgframe.grid(column=0, row=0, sticky=tk.EW)
-        self.tablemaster.grid(column=0, row=2, sticky=tk.NSEW)
-        self.tableframe.grid(column=0, row=0, sticky=tk.NSEW)
-        self.tablemaster.columnconfigure(0, weight=1)
-        self.tablemaster.rowconfigure(0, weight=1)
-        self.tablemaster.rowconfigure(1, weight=1)
-        btnframe.grid(column=0, row=3, sticky=tk.E)
-        for colno, btn_spec in enumerate(button_list):
-            btn_action = ClickSet(self, btn_spec[1]).click
-            btn = ttk.Button(btnframe, text=btn_spec[0], command=btn_action)
-            if btn_spec[2] is not None:
-                self.win.bind(btn_spec[2], btn_action)
-            self.buttons.append(btn)
-            btn.grid(column=colno, row=0, sticky=tk.E, padx=3)
-        # Allow resizing.
-        self.win.columnconfigure(0, weight=1)
-        self.win.rowconfigure(1, weight=0)
-        self.win.rowconfigure(2, weight=2)
-        msgframe.columnconfigure(0, weight=1)
-        btnframe.columnconfigure(0, weight=1)
-        # Other key bindings
-        self.win.protocol("WM_DELETE_WINDOW", self.cancel)
-        # Position window.
-        self.win.update_idletasks()
-        m = re.match(r"(\d+)x(\d+)\+(-?\d+)\+(-?\d+)", self.win.geometry())
-        if m is not None:
-            wwd = int(m.group(1))
-            wht = int(m.group(2))
-            swd = self.win.winfo_screenwidth()
-            sht = self.win.winfo_screenheight()
-            xpos = (swd / 2) - (wwd / 2)
-            ypos = (sht / 2) - (wht / 2)
-            self.win.geometry("%dx%d+%d+%d" % (wwd, wht, xpos, ypos))
-        # Limit resizing
-        self.win.minsize(width=300, height=0)
-
-    def cancel(self: TableUI):
-        """Cancel the dialog."""
-        self.dialog_canceled = True
-        self.win.destroy()
-
-    def activate(self: TableUI):
-        """Activate the dialog."""
-        self.win.grab_set()
-        self.win._root().withdraw()
-        self.win.focus_force()
-        if self.focus_button:
-            self.buttons[self.focus_button].focus()
-        self.win.wait_window(self.win)
-        self.win.update_idletasks()
-        # Explicitly delete the Tkinter variable to suppress Tkinter error message.
-        self.hl_both_var = None
-        rv = self.return_value
-        return (self.button_clicked_value, rv)
-
-
-class ClickSet:
-    """A class for handling button clicks."""
-
-    def __init__(self: ClickSet, ui_obj, button_value: int) -> None:
-        self.ui_obj = ui_obj
-        self.button_value = button_value
-
-    def click(self: ClickSet, *args):
-        """Handle a button click."""
-        self.ui_obj.button_clicked_value = self.button_value
-        self.dialog_canceled = self.button_value is None
-        self.ui_obj.win.destroy()
 
 
 class PgUpsert:
@@ -561,11 +35,9 @@ class PgUpsert:
     To avoid SQL injection, all SQL statements are generated using the [`psycopg2.sql`](https://www.psycopg.org/docs/sql.html) module.
 
     Args:
-        host (str): Name of the PostgreSQL host.
-        database (str): Name of the PostgreSQL database.
-        user (str): Name of the PostgreSQL user. This user must have the necessary permissions to connect to the database, query the information_schema, create temporary objects, select from the staging tables, and update and insert into the base tables. No checking is done to verify these permissions.
-        port (int, optional): PostgreSQL database port, defaults to 5432.
-        passwd (None or str, optional): Password for the PostgreSQL user. If None, the user will be prompted to enter the password. Defaults to None.
+        uri (str or None, optional): Connection URI for the PostgreSQL database. Defaults to None. **Note**: If a connection URI is not provided, an existing connection object must be provided.
+        conn (psycopg2.extensions.connection or None, optional): An existing connection object to the PostgreSQL database. Defaults to None. **Note**: If a connection object is not provided, a connection URI must be provided. If both are provided, the connection object will be used.
+        encoding (str, optional): The encoding to use for the database connection. Defaults to "utf-8".
         tables (list or tuple or None, optional): List of table names to perform QA checks on and upsert. Defaults to ().
         stg_schema (str or None, optional): Name of the staging schema where tables are located which will be used for QA checks and upserts. Tables in the staging schema must have the same name as the tables in the base schema that they will be upserted to. Defaults to None.
         base_schema (str or None, optional): Name of the base schema where tables are located which will be updated or inserted into. Defaults to None.
@@ -2545,299 +2017,3 @@ class PgUpsert:
             self.db.rollback()
         self.db.close()
         return self
-
-
-def treeview_table(
-    parent: ttk.Frame,
-    rowset: list | tuple,
-    column_headers: list | tuple,
-    select_mode="none",
-):
-    """Creates a TreeView table containing the specified data, with scrollbars and
-    status bar in an enclosing frame.
-    This does not grid the table frame in its parent widget. Returns a tuple
-    of 0: the frame containing the table,  and 1: the table widget itself.
-    """
-    nrows = range(len(rowset))
-    ncols = range(len(column_headers))
-    hdrwidths = [len(column_headers[j]) for j in ncols]
-    if len(rowset) > 0:
-        datawidthtbl = [
-            [
-                len(
-                    (rowset[i][j] if isinstance(rowset[i][j], str) else str(rowset[i][j])),
-                )
-                for i in nrows
-            ]
-            for j in ncols
-        ]
-        datawidths = [max(cwidths) for cwidths in datawidthtbl]
-    else:
-        datawidths = hdrwidths
-    colwidths = [max(hdrwidths[i], datawidths[i]) for i in ncols]
-    # Set the font.
-    ff = tkfont.nametofont("TkFixedFont")
-    tblstyle = ttk.Style()
-    tblstyle.configure("tblstyle", font=ff)
-    charpixels = int(1.3 * ff.measure("0"))
-    tableframe = ttk.Frame(master=parent, padding="3 3 3 3")
-    statusframe = ttk.Frame(master=tableframe)
-    # Create and configure the Treeview table widget
-    tv_widget = ttk.Treeview(
-        tableframe,
-        columns=column_headers,
-        selectmode=select_mode,
-        show="headings",
-    )
-    tv_widget.configure()["style"] = tblstyle
-    ysb = ttk.Scrollbar(tableframe, orient="vertical", command=tv_widget.yview)
-    xsb = ttk.Scrollbar(tableframe, orient="horizontal", command=tv_widget.xview)
-    tv_widget.configure(yscrollcommand=ysb.set, xscrollcommand=xsb.set)
-    # Status bar
-    statusbar = ttk.Label(
-        statusframe,
-        text="    %d rows" % len(rowset),
-        relief=tk.RIDGE,
-        anchor=tk.W,
-    )
-    tableframe.statuslabel = statusbar
-    # Fill the Treeview table widget with data
-    set_tv_headers(tv_widget, column_headers, colwidths, charpixels)
-    fill_tv_table(tv_widget, rowset, statusbar)
-    # Place the table
-    tv_widget.grid(column=0, row=0, sticky=tk.NSEW)
-    ysb.grid(column=1, row=0, sticky=tk.NS)
-    xsb.grid(column=0, row=1, sticky=tk.EW)
-    statusframe.grid(column=0, row=3, sticky=tk.EW)
-    tableframe.columnconfigure(0, weight=1)
-    tableframe.rowconfigure(0, weight=1)
-    # Place the status bar
-    statusbar.pack(side=tk.BOTTOM, fill=tk.X)
-    # Allow resizing of the table
-    tableframe.columnconfigure(0, weight=1)
-    tableframe.rowconfigure(0, weight=1)
-    #
-    return tableframe, tv_widget
-
-
-def set_tv_headers(
-    tvtable: ttk.Treeview,
-    column_headers: list,
-    colwidths: list,
-    charpixels: int,
-):
-    """Set the headers and column widths for a Treeview table widget."""
-    pixwidths = [charpixels * col for col in colwidths]
-    for i in range(len(column_headers)):
-        hdr = column_headers[i]
-        tvtable.column(hdr, width=pixwidths[i])
-        tvtable.heading(
-            hdr,
-            text=hdr,
-            command=lambda _col=hdr: treeview_sort_column(tvtable, _col, False),
-        )
-
-
-def treeview_sort_column(tv: ttk.Treeview, col: str, reverse: bool):
-    """Sort a column in a Treeview table widget.
-
-    From https://stackoverflow.com/questions/1966929/tk-treeview-column-sort#1967793
-    """
-    colvals = [(tv.set(k, col), k) for k in tv.get_children()]
-    colvals.sort(reverse=reverse)
-    # Rearrange items in sorted positions
-    for index, (_val, k) in enumerate(colvals):
-        tv.move(k, "", index)
-    # Reverse sort next time
-    tv.heading(col, command=lambda: treeview_sort_column(tv, col, not reverse))
-
-
-def fill_tv_table(tvtable: ttk.Treeview, rowset: list | tuple, status_label=None):
-    """Fill a Treeview table widget with data."""
-    for i, row in enumerate(rowset):
-        enc_row = [c if c is not None else "" for c in row]
-        tvtable.insert(parent="", index="end", iid=str(i), values=enc_row)
-    if status_label is not None:
-        status_label.config(text="    %d rows" % len(rowset))
-
-
-def ellapsed_time(start_time: datetime):
-    """Returns a string representing the ellapsed time since the start time."""
-    dt = (datetime.now() - start_time).total_seconds()
-    if dt < 60:
-        return f"{round((datetime.now() - start_time).total_seconds(), 3)} seconds"
-    if dt < 3600:
-        return f"{int(dt // 60)} minutes, {round(dt % 60, 3)} seconds"
-    return f"{int(dt // 3600)} hours, {int((dt % 3600)) // 60} minutes, {round(dt % 60, 3)} seconds"  # noqa: UP034
-
-
-def clparser() -> argparse.ArgumentParser:
-    """Command line interface for the upsert function."""
-    parser = argparse.ArgumentParser(
-        add_help=False,
-        description=__description__,
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-    )
-    parser.add_argument(
-        "--help",
-        action="help",
-        help="show this help message and exit",
-    )
-    parser.add_argument(
-        "--version",
-        action="version",
-        version=f"%(prog)s {__version__}",
-    )
-    parser.add_argument(
-        "--debug",
-        action="store_true",
-        help="display debug output",
-    )
-    parser.add_argument(
-        "-q",
-        "--quiet",
-        action="store_true",
-        help="suppress all console output",
-    )
-    parser.add_argument(
-        "-l",
-        "--log",
-        type=Path,
-        help="write log to LOGFILE",
-    )
-    parser.add_argument(
-        "-e",
-        "--exclude-columns",
-        dest="exclude",
-        type=str,
-        help="comma-separated list of columns to exclude from null checks",
-    )
-    parser.add_argument(
-        "-n",
-        "--null-columns",
-        dest="null",
-        type=str,
-        help="comma-separated list of columns to exclude from null checks",
-    )
-    parser.add_argument(
-        "-c",
-        "--commit",
-        action="store_true",
-        help="commit changes to database",
-    )
-    parser.add_argument(
-        "-i",
-        "--interactive",
-        action="store_true",
-        help="display interactive GUI of important table information",
-    )
-    parser.add_argument(
-        "-m",
-        "--upsert-method",
-        default="upsert",
-        choices=["upsert", "update", "insert"],
-        help="method to use for upsert",
-    )
-    parser.add_argument(
-        "-h",
-        "--host",
-        required=True,
-        type=str,
-        help="database host",
-    )
-    parser.add_argument(
-        "-p",
-        "--port",
-        required=True,
-        type=int,
-        help="database port",
-    )
-    parser.add_argument(
-        "-d",
-        "--database",
-        required=True,
-        type=str,
-        help="database name",
-    )
-    parser.add_argument(
-        "-u",
-        "--user",
-        required=True,
-        type=str,
-        help="database user",
-    )
-    parser.add_argument(
-        "-s",
-        "--staging-schema",
-        default="staging",
-        dest="stg_schema",
-        required=True,
-        type=str,
-        help="staging schema name",
-    )
-    parser.add_argument(
-        "-b",
-        "--base-schema",
-        default="public",
-        dest="base_schema",
-        required=True,
-        type=str,
-        help="base schema name",
-    )
-    parser.add_argument(
-        "--encoding",
-        default="utf-8",
-        type=str,
-        required=False,
-        help="encoding of the database",
-    )
-    parser.add_argument(
-        "tables",
-        nargs="+",
-        help="table name(s)",
-    )
-    return parser
-
-
-def main() -> None:
-    """Main command line entrypoint for the upsert function."""
-    args = clparser().parse_args()
-    if args.log and args.log.exists():
-        args.log.unlink()
-    if not args.quiet:
-        stream_handler = logging.StreamHandler()
-        stream_handler.setLevel(logging.INFO)
-        stream_handler.setFormatter(logging.Formatter("%(message)s"))
-        logger.addHandler(stream_handler)
-    if args.log:
-        file_handler = logging.FileHandler(args.log)
-        file_handler.setLevel(logging.INFO)
-        logger.addHandler(file_handler)
-    if args.debug:
-        logger.setLevel(logging.DEBUG)
-        formatter = logging.Formatter(
-            "%(asctime)s - %(levelname)s - %(lineno)d - %(message)s",
-        )
-        for handler in logger.handlers:
-            handler.setFormatter(formatter)
-            handler.setLevel(logging.DEBUG)
-    try:
-        PgUpsert(
-            uri=f"postgresql://{args.user}@{args.host}:{args.port}/{args.database}",
-            encoding="utf-8",
-            tables=args.tables,
-            stg_schema=args.stg_schema,
-            base_schema=args.base_schema,
-            do_commit=args.commit,
-            upsert_method=args.upsert_method,
-            interactive=args.interactive,
-            exclude_cols=args.exclude.split(",") if args.exclude else None,
-            exclude_null_check_cols=args.null.split(",") if args.null else None,
-        ).run()
-    except Exception as e:
-        logger.error(e)
-        sys.exit(1)
-
-
-if __name__ == "__main__":
-    main()
